fix: address PR #721 CI failure
